refactor(#9): Event DB 접근을 분석 서비스 API 호출로 변경
