+++ src/main/java/com/modive/adminservice/external/analysis/dto/EventTotalCntByTypeDTO.java
@@ -1,4 +1,4 @@
-package com.modive.adminservice.domain.event.dto;
+package com.modive.adminservice.external.analysis.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class EventsByDriveDTO {
-    private String type;
+public class EventTotalCntByTypeDTO {
+    private String reason;
     private Long count;
 }
