+++ src/main/java/com/modive/adminservice/external/analysis/service/AnalysisService.java
@@ -1,11 +1,11 @@
-package com.modive.adminservice.domain.event.service;
+package com.modive.adminservice.external.analysis.service;
 
 import com.modive.adminservice.api.dashboard.dto.res.TotalEventCntByReasonItem;
-import com.modive.adminservice.domain.event.dto.EventsByDriveDTO;
+import com.modive.adminservice.external.analysis.dto.EventsByDriveDTO;
 
 import java.util.List;
 
-public interface EventService {
+public interface AnalysisService {
     List<EventsByDriveDTO> getTotalEventCntByType(Long driveId);
     List<TotalEventCntByReasonItem> getTotalEventCntByType();
 }
